requirements.txt
numpy>=1.26,<2.0
Pillow>=10.4,<11.0
